crud

+++ app/Controllers/Buku.php
@@ -53,7 +53,7 @@ class Buku extends BaseController
 
         // Proses upload sampul
         $file = $this->request->getFile('a_sampul');
-        $sampulName = $judul_buku . '.' . $file->getExtension(); // Gunakan judul buku sebagai nama file
+        $sampulName = $kode_buku . '.' . $file->getExtension(); // Gunakan judul buku sebagai nama file
         $file->move('assets/img/buku', $sampulName);             // konfigurasi upload
 
         // Data untuk disimpan ke database
@@ -79,28 +79,42 @@ class Buku extends BaseController
 
     public function editBuku()
     {
-        // Ambil data dari form
-        $id = $this->request->getPost('e_idbuku');
-        $data = [
-            'id_buku'     => $this->request->getPost('e_idbuku'),
-            'kode_buku'   => $this->request->getPost('e_kodebuku'),
-            'judul_buku'  => $this->request->getPost('e_judulbuku'),
-            'pengarang'   => $this->request->getPost('e_pengarang'),
-            'penerbit'    => $this->request->getPost('e_penerbit'),
-            'tahun_terbit'=> $this->request->getPost('e_tahunterbit'),
-            'kategori'    => $this->request->getPost('e_kategori'),
-            'no_rak'      => $this->request->getPost('e_norak'),
-            'jumlah_buku' => $this->request->getPost('e_jumlahbuku'),
-        ];
-
-        // Validasi file sampul
-        if ($this->validate([
-            'sampul' => [
+        // Validasi input untuk sampul
+        $validation = $this->validate([
+            'e_sampul' => [
                 'uploaded[e_sampul]',
                 'mime_in[e_sampul,image/jpg,image/jpeg,image/png]',
                 'max_size[e_sampul,2048]', 
-            ]
-        ])) {
+            ],
+        ]);
+
+        // Ambil data dari form
+        $id = $this->request->getPost('e_idbuku');
+        $kode_buku    = $this->request->getPost('e_kodebuku');
+        $judul_buku   = $this->request->getPost('e_judulbuku');
+        $pengarang    = $this->request->getPost('e_pengarang');
+        $penerbit     = $this->request->getPost('e_penerbit');
+        $tahun_terbit = $this->request->getPost('e_tahunterbit');
+        $kategori     = $this->request->getPost('e_kategori');
+        $no_rak       = $this->request->getPost('e_norak');
+        $jumlah_buku  = $this->request->getPost('e_jumlahbuku');
+
+        // Data untuk disimpan ke database
+        $data = [
+            'kode_buku'     => $kode_buku,
+            'judul_buku'    => $judul_buku,
+            'pengarang'     => $pengarang,
+            'penerbit'      => $penerbit,
+            'tahun_terbit'  => $tahun_terbit,
+            'kategori'      => $kategori,
+            'no_rak'        => $no_rak,
+            'jumlah_buku'   => $jumlah_buku,
+        ];
+
+        if (!$validation) {
+            // Tidak ada sampul yang diunggah atau validasi gagal, gunakan sampul lama
+            $data['sampul'] = $this->request->getPost('e_oldsampul');
+        } else {
             // Validasi berhasil, hapus foto sampul lama jika ada
             $old_foto = $this->request->getPost('e_oldsampul');
             if ($old_foto && file_exists('assets/img/buku/' . $old_foto)) {
@@ -109,21 +123,18 @@ class Buku extends BaseController
 
             // Simpan foto sampul baru
             $sampul = $this->request->getFile('e_sampul');
-            $sampul_name = $this->request->getPost('e_kodebuku') . '.' . $sampul->getExtension();
+            $sampul_name = $kode_buku . '.' . $sampul->getExtension();
             $sampul->move('assets/img/buku', $sampul_name);
             $data['sampul'] = $sampul_name;
-        } else {
-            // Validasi gagal, gunakan foto sampul lama
-            $data['sampul'] = $this->request->getPost('e_oldsampul');
-            return redirect()->back()->withInput()
-                                    ->with('errors', 'Maaf, gagal menambahkan data buku. Silakan periksa input Anda.');
         }
 
         // Update data buku di database
         $this->bukuModel->updateBuku($data, $id);
 
+        // Redirect dengan pesan sukses
         return redirect()->to('/admin/buku')->with('message', 'Buku berhasil diubah!');
     }
+
 
 
     public function deleteBuku()
@@ -155,8 +166,6 @@ class Buku extends BaseController
         return redirect()->to('/admin/buku');
     }
     
-
-
 
     // BUKU RUSAK
     public function buku_rusak()
@@ -191,13 +200,13 @@ class Buku extends BaseController
         }
 
         // Ambil data dari request
-        $kode_buku = $this->request->getPost('abkr_kode');
-        $judul_buku = $this->request->getPost('abkr_judul');
-        $pengarang = $this->request->getPost('abkr_pengarang');
-        $kategori = $this->request->getPost('abkr_kategori');
+        $kode_buku         = $this->request->getPost('abkr_kode');
+        $judul_buku        = $this->request->getPost('abkr_judul');
+        $pengarang         = $this->request->getPost('abkr_pengarang');
+        $kategori          = $this->request->getPost('abkr_kategori');
         $tanggal_pendataan = $this->request->getPost('abkr_tgldata');
-        $halaman = $this->request->getPost('abkr_hal');
-        $keterangan = $this->request->getPost('abkr_ket');
+        $halaman           = $this->request->getPost('abkr_hal');
+        $keterangan        = $this->request->getPost('abkr_ket');
         
         // Proses upload foto bukti
         $file = $this->request->getFile('abkr_bukti');
@@ -206,14 +215,14 @@ class Buku extends BaseController
 
         // Data untuk disimpan ke database
         $data = [
-            'kode_buku' => $kode_buku,
-            'judul_buku' => $judul_buku,
-            'pengarang' => $pengarang,
-            'kategori' => $kategori,
+            'kode_buku'         => $kode_buku,
+            'judul_buku'        => $judul_buku,
+            'pengarang'         => $pengarang,
+            'kategori'          => $kategori,
             'tanggal_pendataan' => $tanggal_pendataan,
-            'halaman' => $halaman,
-            'keterangan' => $keterangan,
-            'foto_bukti' => $buktiName,
+            'halaman'           => $halaman,
+            'keterangan'        => $keterangan,
+            'foto_bukti'        => $buktiName,
         ];
 
         // Simpan data menggunakan model
@@ -225,48 +234,60 @@ class Buku extends BaseController
 
     public function editBkr()
     {
-        // Ambil data dari form
-        $id = $this->request->getPost('ebkr_id');
-        $data = [
-            'id_buku'     => $this->request->getPost('ebkr_id'),
-            'kode_buku'   => $this->request->getPost('ebkr_kode'),
-            'judul_buku'  => $this->request->getPost('ebkr_judul'),
-            'pengarang'   => $this->request->getPost('ebkr_pengarang'),
-            'tanggal_pendataan'    => $this->request->getPost('ebkr_tgl'),
-            'keterangan'=> $this->request->getPost('ebkr_ket'),
-        ];
-
-        // Validasi file sampul
-        if ($this->validate([
-            'foto_bukti' => [
+        // Validasi input untuk foto bukti
+        $validation = $this->validate([
+            'ebkr_bukti' => [
                 'uploaded[ebkr_bukti]',
                 'mime_in[ebkr_bukti,image/jpg,image/jpeg,image/png]',
-                'max_size[ebkr_bukti,2048]', 
-            ]
-        ])) {
-            // Validasi berhasil, hapus foto sampul lama jika ada
+                'max_size[ebkr_bukti,2048]',
+            ],
+        ]);
+
+        // Ambil data dari form
+        $id                = $this->request->getPost('ebkr_id');
+        $kode_buku         = $this->request->getPost('ebkr_kode');
+        $judul_buku        = $this->request->getPost('ebkr_judul');
+        $pengarang         = $this->request->getPost('ebkr_pengarang');
+        $kategori          = $this->request->getPost('ebkr_kategori');
+        $tanggal_pendataan = $this->request->getPost('ebkr_tgl');
+        $halaman           = $this->request->getPost('ebkr_hal');
+        $keterangan        = $this->request->getPost('ebkr_ket');
+
+        // Data untuk disimpan ke database
+        $data = [
+            'kode_buku'         => $kode_buku,
+            'judul_buku'        => $judul_buku,
+            'pengarang'         => $pengarang,
+            'kategori'          => $kategori,
+            'tanggal_pendataan' => $tanggal_pendataan,
+            'halaman'           => $halaman,
+            'keterangan'        => $keterangan,
+        ];
+
+        if (!$validation) {
+            // Tidak ada bukti yang diunggah atau validasi gagal, gunakan bukti lama
+            $data['foto_bukti'] = $this->request->getPost('ebkr_oldbukti');
+        } else {
+            // Validasi berhasil, hapus foto bukti lama jika ada
             $old_foto = $this->request->getPost('ebkr_oldbukti');
             if ($old_foto && file_exists('assets/img/bukti/' . $old_foto)) {
                 unlink('assets/img/bukti/' . $old_foto);
             }
 
-            // Simpan foto sampul baru
-            $sampul = $this->request->getFile('ebkr_bukti');
-            $sampul_name = $this->request->getPost('ebkr_kode') . '.' . $sampul->getExtension();
-            $sampul->move('assets/img/bukti', $sampul_name);
-            $data['foto_bukti'] = $sampul_name;
-        } else {
-            // Validasi gagal, gunakan foto sampul lama
-            $data['foto_bukti'] = $this->request->getPost('ebkr_oldbukti');
-            return redirect()->back()->withInput()
-                                    ->with('errors', 'Maaf, gagal menambahkan data buku. Silakan periksa input Anda.');
+            // Simpan foto bukti baru
+            $bukti = $this->request->getFile('ebkr_bukti');
+            $buktiName = $kode_buku . '.' . $bukti->getExtension();
+            $bukti->move('assets/img/bukti', $buktiName);
+            $data['foto_bukti'] = $buktiName;
         }
 
         // Update data buku di database
         $this->bukuModel->updateBkr($data, $id);
 
+        // Redirect dengan pesan sukses
         return redirect()->to('/admin/bukuRusak')->with('message', 'Buku rusak berhasil diubah!');
     }
+
 
     public function deleteBkr()
     {
